% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_local_job.R
\name{as_local_job}
\alias{as_local_job}
\title{Run some R code as a local job}
\usage{
as_local_job(code, name = "Local job", workingDir = ".", importEnv = TRUE, ...)
}
\arguments{
\item{code}{Code to run in a new session}

\item{name, workingDir, importEnv, ...}{Passed to \code{rstudioapi::jobRunScript()}}
}
\value{
The location of the temporary R script created for the job
}
\description{
This lets you execute long-running code in a new session, leaving you free
to continue working while your code runs. See
\href{https://solutions.rstudio.com/r/jobs/}{the RStudio documentation} for more
information.
}
\details{
This function works by copying the supplied code into a temporary R script
and then using \code{rstudioapi::jobRunScript()} to run it as a local job.
}
