% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Divide a numeric vector into bins}
\usage{
bin(
  x,
  width,
  boundary = 0,
  scale = NULL,
  min = -Inf,
  max = Inf,
  format = NULL,
  format_smaller = format,
  format_bigger = format,
  equality_side = NULL,
  reverse = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{width}{The width for each bin}

\item{boundary}{A boundary on which to change bin}

\item{scale}{Currently unused}

\item{min, max}{Values which fall outside of these ranges will be grouped
together}

\item{format}{Can be either:
\itemize{
\item A vectorised function taking two inputs (the lower and upper bounds of the
bin) and returning a character vector
\item A glue-string using the special values \code{x} and \code{y}, e.g. \code{"[{x}, {y})"}
}}

\item{format_smaller, format_bigger}{Can be used to provide special formatting
for number smaller/bigger than min/max.}

\item{equality_side}{If \code{NULL}:
\itemize{
\item Intervals will be closed on the left side if \code{format} is a string
beginning with \code{[}
\item Intervals will be closed on the right side
Otherwise, specify \code{"left"}/\code{"right"} to for intervals closed on the
left/right
}}

\item{reverse}{If \code{TRUE}, factors will be ordered largest to smallest}
}
\value{
A factor
}
\description{
Divide a numeric vector into bins
}
\examples{
comma <- function(x) format(x, big.mark = ",", scientific = FALSE, justify = "none")

bin(
  c(-100000, -3300, 1, 10, 40, 9030), 5000, 
  format = "{comma(x)} to {comma(y)}", 
  format_smaller = "Less than {comma(y)}",
  format_bigger = "Bigger than {comma(x)}",
  min = -20, max = 4000
)
}
