% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils4r-package.R
\docType{package}
\name{utils4r-package}
\alias{utils4r}
\alias{utils4r-package}
\title{utils4r: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
